Updating maingame.py (positioning issues)
Although code does run, the images (background and ground) are not rendered in the correct positions.  The background does not fill the whole window, so there is a horizontal black bar across the bottom of the window. Also, the ground with grass and soil is rendered at the top left, which is not what I want.