Update maingame.py (positioning issues)
The scenery png file is 723x300 and my window is larger in terms of width and height so I need to stretch the image vertically. Here, I used the pygame.transform.scale() method from the documentation to achieve this. The background scenery still has a center point of the middle of the window.