Update maingame.py (positioning issues)
Next, I went into the Floor class and changed the Y coordinate so it will be re-positioned near the bottom at (0,420). The center of the ground has its dimensions centered slightly differently to account for its new top-left positioning; the center is now at (355,440). I found the suitable position by trial and error.