Update maingame.py
I made use of the 2D vector defined at the beginning to add vertical and horizontal vectors for position, velocity, and acceleration. The sprite for my Hero (character) was loaded into pygame. The direction of the character will be initially facing right. An object of Hero class was created and the sprite has been blitted in the game while loop. The sprite is located at the top-left (0,0), but this will be positioned to the floor later.