Update maingame.py
The next task is after rendering the character in the pygame window is to allow for its the sprite's movement via the move_H Hero class method. I have added a y-component to the acceleration vector which applies a sense of gravity, along with whether the sprite is running or not based off of the Boolean value of the run member variable and a default value of 0.3. The sprite is only allowed to move left and right not up and down when the arrow keys (jumping will be dealt with elsewhere). 
Moving left: the x-component of the acceleration will be negative.
Moving right: the x-component of the acceleration will be positive.
I used the width of the pygame screen, sprite width and velocity to restrict movement of the character on the edges of the window.