Update maingame.py
Next formulas are added to change the acceleration, velocity and position of the sprite. Friction acts as a resistance to slow down the acceleration in the for the x-component. The velocity is updated with acceleration vector that helps to slow down the character's speed in the x-direction so the Hero sprite does not move really fast. The position vector is finally updated with acceleration and velocity.  The position (self.position_H) is matched to the top-left rectangle of the sprite.