Update maingame.py
Next there is needs to be some collision detection put between the sprite and ground (floor). The acceleration vector in the move_H Hero method will cause the sprite to fall off the bottom of the window so there the rectangle for the floor and the sprite have to compared. So the floor object is put into its own group outside of the game loop. Then inside of the game loop, there is a check to see if the velocity for the sprite is above zero(0) meaning the character is moving downwards; the lowest_point variable stores the collision of the hero object and the floor _group. If the sprite position (the y-component of the hero position) is just above its surrounding rectangle at the point of collision (rect.bottom) there is a small correction to move the sprite up 1 pixel to act as a slowdown factor against acceleration and velocity. Although this stage is not about jumping I added a variable isJumping to the Hero class constructor as limiter because when the character and the floor are in contact jumping should not be possible  (False) unless the space bar is pressed down on.