Update maingame.py
Next a new member variable is added to the Hero class called jumpCount. This jumpCount is to be used as part of a quadratic formula to simulate the jumping of an inverted parabola in the jump_H Hero function. The rectangle of the Herp sprite is moved 1 pixel to keep the alignment of the hitbox around the sprite. If isJumping is False then isJumping is set to True and the velocity increases upwards (in the negative direction by -10). If isJumping is True  then the else conditional block is executed. Inside the else block is a nested if-else conditional statement to check whether the jumpCount set in the Hero constructor is equal to or greater than -10. If it is the quadratic formula on line 133 is subtracted from the y-component of the position vector of the sprite.  The negVal variable is used to cause the character to fall back down to the ground when peak height in jump is reached, velocity is zero (0) and the jumpCount value becomes negative. The jumpCount is reduced by 1. When -10 in the jumpCount is reached jumpCount is reset for the next future jump.