Update maingame.py
Now the in the game loop I will add the conditional statement for tracking the key events. So when the user presses the spacebar the jump_H method will be called from under the pygame.event.get() method. But in the current state the sprite/character can jump infinitely even when in the air and move towards the top of the pygame window; to prevent this on line 175 I added a check to make sure that the hero is not jumping (isJumping =False) before the jump_H method is called so another jump can only be executed when the sprite is on ground again. I cleaned up some areas of the code too.