Update maingame.py
I decided to take away the member variable for the Hero class called self.direc_H and instead use left and right member variables for the class so when the sprite is moving left and right it can be tracked by Boolean value instead of a string "RIGHT" and "LEFT" because there are states when the sprite is standing still and not moving left or right. Also, I put in the class method updateF_H for Hero class and two lists for animating the character running right and left. So, these would be run_anim_L and and run_anim_R. Each list carries 7 images that will cycle in the game loop as the character moves left and right.
For the updateF_H method a variable mv_frame is used to make he end of the animations in those two aforementioned lists. There are 7 right and 7 left running frames but index will start from 0. If the sprite is not jumping and is running then a velocity check of the x-component is done. If it's greater than 0 the sprite is assumed to be moving right and image of the particular right frame is list is rendered with the appropriate mv_frame index. Otherwise, the sprite is assumed to be running left and Boolean values for right and left are set in the opposite direction. Outside the if-else conditional statement the mv_frame is incremented by 1. If the sprite slows down below 0.2 (value for velocity_H) in the x-direction then the mv_frame is reset to zero (0).