Update maingame.py
I added a method for the attacking Hero Sprite called attack_H. If we go over index 10 (11 attack frames) (denoted by the a_Frame variable) then the a_Frame variable is reset to zero. I have also loaded the frame .png files for attacking left and right; some have been repeated as the sword swipe is very quick, so see it adequately some attacks frames are repeated.  In the attack_H class method, I noted that if the sprite is facing right the right attack animations run if not the ones for left attack run. Then the frame is incremented by 1 at the end.  I have added a  turn_Correction class method for when the character is facing left because I found positioning was odd when I ran the program.