Update maingame.py
I noticed that in the transition from walking animation into attack animation the center of the hero (character) sprite is offset and that makes the attack look odd so I adjusted the position of the character when it is doing a horizontal attack. The x-component of the position vector will get moved forward 20 pixels (added) when going from walking to attack and the 20 pixels will taken off (subtracted) when moving from attack back to walking. This applies to the second frame and the last frame in the attack animation lists.