Update maingame.py
I changed the pixel correction down moderately for the turn_Correction method of the Hero class to +- 10.  In the game (while) loop I put the attack animation to key A from line 250. There is a guard to make sure the hero is only attacking when attacking member variable is set to False. then after the attack_H method is called the attacking variable is set to a Boolean value of True. But the guard only applies when a key is pressed down (with the conditional if statement). The Boolean guard has to be applied outside the if statement (if event.key == pygame.K_a:) below the drawGameWindow() so index in the left and right attack frames can be moved forwards.